[WEB-3711] add custom trait
